Punchlist items
Made the menu items point to stuff about me.
Add trailing slashes to the API URLs so that I didn’t do multiple GET’s
to load a page - this was breaking the CORS support since the first one
was a 301.
Document why I’m using a relative path in the Javascript Ajax call.