Refactored the code to assist with testing
Make the outer object always have the same keys, and encapsulate the
specific data for each API within one of those keys, so the API return
data structure is more consistent. Added helper APIs to keep from
modifying instance variables directly. Make it so even in a failure of
the URL parse, the returned object still conforms to the spec.